refactor: reorganize pytmx into modular components with improved type hints
