+++ pytmx/util_pygame_sdl2.py
@@ -56,9 +56,9 @@ def handle_flags(flags: Optional[pytmx.TileFlags]) -> tuple[float, bool, bool]:
 
     if flags.flipped_diagonally:
         if flags.flipped_vertically:
-            return 270, False, False
+            return 270.0, False, False
         else:
-            return 90, False, False
+            return 90.0, False, False
     else:
         return 0.0, flags.flipped_horizontally, flags.flipped_vertically
